Se encapsula el carrito en la vista de tienda.
